requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
kafka-python==2.0.2
mysql-connector-python==9.0.0
requests==2.32.3
six==1.16.0
sseclient==0.0.27
urllib3==2.2.2